Lighthouse fix (#1)
* Add title element

LIghthouse flagged an SEO best practice that I was missing. Adding this property
to the config and checking for it in the web browser inspector is how I fixed it

* Add descriptions to posts

LIghthouse flagged an SEO best practice that I was missing. Adding this front
matter updated the `property="og:description"` meta tags, but I don't know if
that's the same thing

* Update links in `docker-practice.md`

* Update links in `clojurescript-project-setup.md`

* Add a default description